update cru accessory

+++ src/main/java/com/colorbike/dto/Accessory.java
@@ -1,7 +1,7 @@
-
 package com.colorbike.dto;
 
 import java.io.Serializable;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,11 +13,12 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
-public class Accessory implements Serializable{
+public class Accessory implements Serializable {
+
     private int accessoryId;
     private String accessoryName;
     private String accessoryImage;
     private String accessoryImageicon;
     private String accessoryDescription;
-    private int price;
+    private double price;
 }
